dedup some stuff on the local and spark backends
