+++ hail/python/hail/backend/py4j_backend.py
@@ -135,6 +135,24 @@ class Py4JBackend(Backend):
     def _remove_reference_from_scala_backend(self, name):
         self._jbackend.pyRemoveReference(name)
 
+    def from_fasta_file(self, name, fasta_file, index_file, x_contigs, y_contigs, mt_contigs, par):
+        return json.loads(self._jbackend.pyFromFASTAFile(name, fasta_file, index_file, x_contigs, y_contigs, mt_contigs, par))
+
+    def load_references_from_dataset(self, path):
+        return json.loads(self._jbackend.pyLoadReferencesFromDataset(path))
+
+    def add_sequence(self, name, fasta_file, index_file):
+        self._jbackend.pyAddSequence(name, fasta_file, index_file)
+
+    def remove_sequence(self, name):
+        self._jbackend.pyRemoveSequence(name)
+
+    def add_liftover(self, name, chain_file, dest_reference_genome):
+        self._jbackend.pyAddLiftover(name, chain_file, dest_reference_genome)
+
+    def remove_liftover(self, name, dest_reference_genome):
+        self._jbackend.pyRemoveLiftover(name, dest_reference_genome)
+
     @property
     def requires_lowering(self):
         return True
